refactor: Refactor run
