+++ src/lib.rs
@@ -5,3 +5,6 @@ pub mod error;
 pub mod extractor;
 pub mod server;
 pub mod typst_lib;
+pub mod run;
+
+pub use run::run;
